#23: Moved injection test into separate class
