+++ src/test/java/systems/composable/dropwizard/cassandra/DropwizardCassandraIntegrationTest.java
@@ -21,7 +21,6 @@ import com.google.common.io.Resources;
 import io.dropwizard.testing.junit.DropwizardAppRule;
 import org.junit.ClassRule;
 import org.junit.Test;
-import systems.composable.dropwizard.cassandra.smoke.SmokeInjectedApp;
 import systems.composable.dropwizard.cassandra.smoke.SmokeTestApp;
 import systems.composable.dropwizard.cassandra.smoke.SmokeTestConfiguration;
 
@@ -29,7 +28,6 @@ import javax.ws.rs.client.ClientBuilder;
 import javax.ws.rs.client.WebTarget;
 import javax.ws.rs.core.UriBuilder;
 import java.net.URI;
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -39,46 +37,17 @@ public class DropwizardCassandraIntegrationTest {
     public static final DropwizardAppRule<SmokeTestConfiguration> APP =
         new DropwizardAppRule<>(SmokeTestApp.class, Resources.getResource("minimal.yml").getPath());
 
-    @ClassRule
-    public static final DropwizardAppRule<SmokeTestConfiguration> APP_INJ =
-        new DropwizardAppRule<>(SmokeInjectedApp.class, Resources.getResource("injected.yml").getPath());
-
-    private void canQuery(URI uri) {
-        final WebTarget target = ClientBuilder.newClient().target(uri);
-        final List<String> result = Lists.newArrayList(target.request().get(String[].class));
-        assertThat(result).contains("system");
-    }
-
     @Test
     public void canQueryCassandra() throws Exception {
-        canQuery(UriBuilder.fromUri("http://localhost")
-            .port(APP.getLocalPort())
-            .path("query")
-            .build());
-    }
-
-    @Test
-    public void canQuerySessionField() throws Exception {
-        canQuery(UriBuilder.fromUri("http://localhost")
-            .port(APP_INJ.getLocalPort())
-            .path("querySessionField")
-            .build());
-    }
-
-    @Test
-    public void canQuerySessionParameter() throws Exception {
-        canQuery(UriBuilder.fromUri("http://localhost")
-            .port(APP_INJ.getLocalPort())
-            .path("querySessionParameter")
-            .build());
-    }
-
-    @Test
-    public void canQueryClusterField() throws Exception {
-        canQuery(UriBuilder.fromUri("http://localhost")
-            .port(APP_INJ.getLocalPort())
-            .path("queryClusterField")
-            .build());
+        assertThat(Lists.newArrayList(
+            ClientBuilder.newClient()
+                .target(UriBuilder.fromUri("http://localhost")
+                    .port(APP.getLocalPort())
+                    .path("query")
+                    .build())
+                .request()
+                .get(String[].class)))
+            .contains("system");
     }
 
     @Test
